handle ssh things and cleanup
