+++ ui/app/templates/vault/cluster/secrets/backend/configuration.hbs
@@ -15,7 +15,7 @@
 />
 
 {{#if this.isConfigurable}}
-  {{! only relevant to SSH and AWS, eventually Azure and GCP }}
+  {{! only configurable secret engines are SSH and AWS. Eventually we will add Azure and GCP }}
   <Toolbar>
     <ToolbarActions>
       <ToolbarLink
@@ -28,7 +28,7 @@
     </ToolbarActions>
   </Toolbar>
 
-  <SecretEngine::ConfigurableSecretEngineDetails @model={{this.model}} @isConfigurable={{this.isConfigurable}} />
+  <SecretEngine::ConfigurableSecretEngineDetails @model={{this.model}} />
 
   <SecretsEngineMountConfig @model={{this.model}} class="has-top-margin-xl has-bottom-margin-xl" data-test-mount-config />
